requirements.txt
fastapi
pydantic
python-multipart
python-dotenv